chore: reorganize components into layout and home folders
