+++ src/layout/MainLayout.jsx
@@ -1,4 +1,4 @@
-import Navbar from '../components/Navbar';
+import Navbar from '../components/layout/Navbar';
 
 function MainLayout({ children }) {
   return (
